+++ UnBBayes/src/unbbayes/aprendizagem/CL.java
@@ -18,7 +18,7 @@ public class CL {
 	public int classe;
 	public int caseNumber;
 	public int raiz;
-	protected byte[][] dataBase;
+	protected int[][] dataBase;
     protected int[] vector;
     protected boolean compacted;
     protected double [][] matrizinfo;
@@ -39,7 +39,7 @@ public class CL {
 	/*
 	 * Prepara e faz tudo.
 	 */
-	public void preparar(NodeList vetordevariaveis, int classei, int numerodecasos, int[] vetor,boolean comp,byte[][]dados) {
+	public void preparar(NodeList vetordevariaveis, int classei, int numerodecasos, int[] vetor,boolean comp,int[][]dados) {
 		int i;
 		//copiar os parametros
 		nvar=vetordevariaveis.size();
